+++ src/scl/utility/hash.h
@@ -8,6 +8,7 @@
 
 #include <scl/utility/hash/byte_view.h>
 #include <scl/utility/hash/concepts.h>
+#include <scl/utility/hash/constant_bytes.h>
 #include <scl/utility/hash/djb2.h>
 #include <scl/utility/hash/fnv1a.h>
 #include <scl/utility/hash/jenkins_ota.h>